# LMDB optimizations