--- a/ocpmodels/datasets/single_point_lmdb.py
+++ b/ocpmodels/datasets/single_point_lmdb.py
@@ -35,21 +35,19 @@ class SinglePointLmdbDataset(Dataset):
             self.db_path
         )
 
-        env = self.connect_db(self.db_path)
+        self.env = self.connect_db(self.db_path)
 
         self._keys = [
-            f"{j}".encode("ascii") for j in range(env.stat()["entries"])
+            f"{j}".encode("ascii") for j in range(self.env.stat()["entries"])
         ]
         self.transform = transform
-
-        env.close()
 
     def __len__(self):
         return len(self._keys)
 
     def __getitem__(self, idx):
         # Return features.
-        env = self.connect_db(self.db_path)
+        env = self.env
         datapoint_pickled = env.begin().get(self._keys[idx])
         data_object = pickle.loads(datapoint_pickled)
         data_object = (
@@ -57,7 +55,6 @@ class SinglePointLmdbDataset(Dataset):
             if self.transform is None
             else self.transform(data_object)
         )
-        env.close()
 
         return data_object
 
@@ -68,6 +65,7 @@ class SinglePointLmdbDataset(Dataset):
             readonly=True,
             lock=False,
             readahead=False,
-            map_size=1099511627776 * 2,
+            meminit=False,
+            max_readers=1,
         )
         return env