--- a/ocpmodels/trainers/energy_trainer.py
+++ b/ocpmodels/trainers/energy_trainer.py
@@ -274,6 +274,12 @@ class EnergyTrainer(BaseTrainer):
             else:
                 self.save(epoch + 1, self.metrics)
 
+        self.train_dataset.close_db()
+        if "val_dataset" in self.config:
+            self.val_dataset.close_db()
+        if "test_dataset" in self.config:
+            self.test_dataset.close_db()
+
     def _forward(self, batch_list):
         output = self.model(batch_list)