--- a/ocpmodels/trainers/forces_trainer.py
+++ b/ocpmodels/trainers/forces_trainer.py
@@ -431,6 +431,12 @@ class ForcesTrainer(BaseTrainer):
                 else:
                     self.save(epoch + 1, self.metrics)
 
+        self.train_dataset.close_db()
+        if "val_dataset" in self.config:
+            self.val_dataset.close_db()
+        if "test_dataset" in self.config:
+            self.test_dataset.close_db()
+
     def _forward(self, batch_list):
         # forward pass.
         if self.config["model_attributes"].get("regress_forces", True):