Some LMDB optimizations following discussions/guidance from the following. Mainly leaving the lmdb open after each get item call.

https://github.com/pytorch/vision/blob/4076a54a67e0d0be94352a3360e364c63677a8ec/torchvision/datasets/lsun.py#L22-L23
https://discuss.pytorch.org/t/whats-the-best-way-to-load-large-data/2977/

- [x] Verify `max_readers=1` compatibility with DDP