@mshuaibii 2020-12-12
> Changes LGTM! Do you have a speed comparison b/w the older and newer dataloader to load say 10k batches?

I'll get those results and update it here