@mshuaibii 2020-12-12
> Looks good.
> Did you test this with the full dataset and hundreds of simultaneous readers? There could be problems that show up when an LMDB file is opened by many readers at the same time.

Good point - I'll give this a test. I only tried it with DDP but that was at most 4 simultaneous readers.