@mshuaibii 2020-12-14
> Changes LGTM! Do you have a speed comparison b/w the older and newer dataloader to load say 10k batches?

As far as indexing the dataset is concerned - indexing the first 10k objects:
```
IS2RE:
Old - 16.2s
New - 3.44s
```
```
S2EF
Old - 15.34
New - 2.83s
```

More practically, dataloader statistics:
10k batches `(batch_size=32, num_workers=32, pin_memory=True)`
```
IS2RE
Old loader: 401.98
New loader: 265.4s
```
```
S2EF
Old loader: 392.02s
New loader: 366.2s
```
It would be useful if one of you could benchmark this on your end please. Slight concerns with the file system on my end that's making it hard to get consistent statistics here.

@abhshkdz @anuroopsriram