@mshuaibii 2020-12-14
> Looks good.
> Did you test this with the full dataset and hundreds of simultaneous readers? There could be problems that show up when an LMDB file is opened by many readers at the same time.

Tested this out with 1k readers on both datasets, no issues.