@mshuaibii 2020-12-21
> Dataloader stats at my end:
> 
> S2EF -- loading 10k batches with batch_size=32 and num_workers=32.
> 
> ```
> Old: 215.38s
> New: 207.12s
> ```
> 
> IS2RE -- loading 10k batches with batch_size=32 and num_workers=32:
> 
> ```
> Old: 45.24s
> New: 40.89s
> ```
> 
> Seeing improvements at my end as well; so good to merge!
> 
> Not necessarily concerning this PR — the time difference in loading 10k `SinglePointLmdbDataset` batches at your vs. my end is significant @mshuaibii. Could you re-run that please? Here's my script for reference:
> 
> ```
> import time
> from tqdm import tqdm
> from torch.utils.data import DataLoader
> from ocpmodels.datasets import SinglePointLmdbDataset, data_list_collater
> 
> 
> if __name__ == "__main__":
>     dataset_config = {
>         "src": "/checkpoint/electrocatalysis/relaxations/features/init_to_relaxed/all/train/data.lmdb",
>     }
>     dataset = SinglePointLmdbDataset(dataset_config)
> 
>     data_loader = DataLoader(
>         dataset,
>         batch_size=32,
>         shuffle=False,
>         collate_fn=data_list_collater,
>         num_workers=32,
>         pin_memory=True,
>     )
> 
>     for i, batch in tqdm(enumerate(data_loader), total=len(data_loader)):
>         if i > 10000:
>             break
> ```

Very marginal improvements as far as this PR is concerned. Interesting your numbers look significantly better...Reran it on my end:

IS2RE
```
Old - 313s
New - 306s
```

I want to say it's a difference in our file systems/network but our S2EF numbers are not wildly off, reran that with times of ~270s. Maybe something about having multiple LMDBs in the S2EF  datasets makes accessing those the bottleneck (hence comparable numbers) versus indexing a single LMDB as in the IS2RE case?