@mshuaibii 2020-12-21
Merging this for now. The benchmarks we ran here however suggest that there could be room for improvement in the `TrajectoryLmdbDataset`. The number of LMDB chunks seems to impact the dataloading speeds. Will leave for a future PR.